core/requirements.txt
absl-py==0.12.0
altgraph==0.17.2
appdirs==1.4.4
astunparse==1.6.3
attrs==21.2.0
audioread==2.1.9
autopep8==1.5.7
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.5
colorama==0.4.4
conda==4.10.3
cycler==0.11.0
Cython==0.29.24
dataclasses==0.6
decorator==5.1.0
dill==0.3.4
dm-tree==0.1.6
fire==0.4.0
flatbuffers==1.12
fonttools==4.28.2
future==0.18.2
gast==0.4.0
gin-config==0.5.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.30.1
google-cloud-core==2.2.1
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
grpcio-tools==1.42.0
h5py==3.6.0
httplib2==0.20.2
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.4.0
joblib==1.1.0
kaggle==1.5.12
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
librosa==0.8.1
llvmlite==0.37.0
lxml==4.6.4
Markdown==3.3.4
matplotlib==3.5.0
menuinst==1.4.16
MouseInfo==0.1.3
mss==6.1.0
neural-structured-learning==1.3.1
numba==0.54.1
numpy==1.20.3
oauthlib==3.1.1
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pefile==2021.9.3
Pillow==8.4.0
pooch==1.5.2
promise==2.3
proto-plus==1.19.8
protobuf==3.19.1
psutil==5.8.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pybind11==2.8.1
pycodestyle==2.7.0
pycosat==0.6.3
PyGetWindow==0.0.9
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.3
PyMsgBox==1.0.9
pyparsing==3.0.6
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.2
python-slugify==5.0.2
pytweening==1.0.4
pytz==2021.3
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
sentencepiece==0.1.96
setuptools-scm==6.3.2
six==1.16.0
SoundFile==0.10.3.post1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-addons==0.15.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.9.0
tensorflow-io-gcs-filesystem==0.22.0
tensorflow-metadata==1.4.0
tensorflow-model-optimization==0.7.0
tensorflowjs==3.11.0
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.3.0
tf-slim==1.1.0
tflite-model-maker==0.3.2
tflite-support==0.3.0
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
typeguard==2.13.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.25.11
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.13.3
zipp==3.5.0
